pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "monad-core"
version = "0.5.0"
description = "Personal AGI Operating Core"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Yam", email = "hscspring@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai>=1.0.0",
    "loguru>=0.7.0",
    "scrapling>=0.1.0",
    "playwright>=1.40.0",
    "curl_cffi>=0.7.0,<0.14.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "websockets>=11.0.3",
    "jieba>=0.42"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0"
]
feishu = [
    "lark-oapi>=1.0.0"
]
desktop = [
    "mss>=9.0.0",
    "pynput>=1.7.0",
    "rapidocr-onnxruntime>=1.3.0",
]
all = [
    "monad-core[feishu,desktop]",
]

[project.scripts]
monad = "monad.main:main"

[tool.setuptools]
packages = ["monad", "monad.cognition", "monad.core", "monad.execution", "monad.interface", "monad.knowledge", "monad.learning", "monad.tools"]

[tool.setuptools.package-data]
"monad.interface" = ["static/*"]
"monad" = ["knowledge/**/*", "knowledge/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: tests that import scrapling/playwright (slow)",
]
